refactor(backend): extract shared REST auth resolver and mappers
