+++ backend/src/main/kotlin/com/travelcompanion/interfaces/rest/ItineraryController.kt
@@ -1,15 +1,11 @@
 package com.travelcompanion.interfaces.rest
 
 import com.travelcompanion.application.trip.ItineraryV2Service
-import com.travelcompanion.domain.trip.Trip
 import com.travelcompanion.domain.trip.TripId
-import com.travelcompanion.domain.user.UserId
-import com.travelcompanion.interfaces.rest.dto.DayContainerResponse
 import com.travelcompanion.interfaces.rest.dto.ItineraryItemV2Request
-import com.travelcompanion.interfaces.rest.dto.ItineraryItemV2Response
-import com.travelcompanion.interfaces.rest.dto.ItineraryV2Response
-import com.travelcompanion.interfaces.rest.dto.ItemContainerResponse
 import com.travelcompanion.interfaces.rest.dto.MoveItineraryItemV2Request
+import com.travelcompanion.interfaces.rest.support.AuthPrincipalResolver
+import com.travelcompanion.interfaces.rest.support.ItineraryResponseMapper
 import jakarta.validation.Valid
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
@@ -30,16 +26,18 @@ import org.springframework.web.bind.annotation.RestController
 @RequestMapping("/trips/{tripId}/itinerary")
 class ItineraryController(
     private val itineraryV2Service: ItineraryV2Service,
+    private val authPrincipalResolver: AuthPrincipalResolver,
 ) {
     @GetMapping("/v2")
     fun getV2(
         authentication: Authentication,
         @PathVariable tripId: String,
     ): ResponseEntity<Any> {
-        val userId = requireUserId(authentication) ?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build()
+        val userId = authPrincipalResolver.userId(authentication)
+            ?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build()
         val id = TripId.fromString(tripId) ?: return ResponseEntity.status(HttpStatus.BAD_REQUEST).build()
         val trip = itineraryV2Service.get(id, userId) ?: return ResponseEntity.status(HttpStatus.NOT_FOUND).build()
-        return ResponseEntity.ok(toItineraryV2Response(trip))
+        return ResponseEntity.ok(ItineraryResponseMapper.toV2Response(trip))
     }
 
     @PostMapping("/v2/items")
@@ -48,7 +46,8 @@ class ItineraryController(
         @PathVariable tripId: String,
         @Valid @RequestBody request: ItineraryItemV2Request,
     ): ResponseEntity<Any> {
-        val userId = requireUserId(authentication) ?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build()
+        val userId = authPrincipalResolver.userId(authentication)
+            ?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build()
         val id = TripId.fromString(tripId) ?: return ResponseEntity.status(HttpStatus.BAD_REQUEST).build()
         val trip = itineraryV2Service.addItem(
             tripId = id,
@@ -59,7 +58,7 @@ class ItineraryController(
             longitude = request.longitude,
             dayNumber = request.dayNumber,
         ) ?: return ResponseEntity.status(HttpStatus.NOT_FOUND).build()
-        return ResponseEntity.status(HttpStatus.CREATED).body(toItineraryV2Response(trip))
+        return ResponseEntity.status(HttpStatus.CREATED).body(ItineraryResponseMapper.toV2Response(trip))
     }
 
     @PutMapping("/v2/items/{itemId}")
@@ -69,7 +68,8 @@ class ItineraryController(
         @PathVariable itemId: String,
         @Valid @RequestBody request: ItineraryItemV2Request,
     ): ResponseEntity<Any> {
-        val userId = requireUserId(authentication) ?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build()
+        val userId = authPrincipalResolver.userId(authentication)
+            ?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build()
         val id = TripId.fromString(tripId) ?: return ResponseEntity.status(HttpStatus.BAD_REQUEST).build()
         val trip = itineraryV2Service.updateItem(
             tripId = id,
@@ -81,7 +81,7 @@ class ItineraryController(
             longitude = request.longitude,
             dayNumber = request.dayNumber,
         ) ?: return ResponseEntity.status(HttpStatus.NOT_FOUND).build()
-        return ResponseEntity.ok(toItineraryV2Response(trip))
+        return ResponseEntity.ok(ItineraryResponseMapper.toV2Response(trip))
     }
 
     @PostMapping("/v2/items/{itemId}/move")
@@ -91,7 +91,8 @@ class ItineraryController(
         @PathVariable itemId: String,
         @Valid @RequestBody request: MoveItineraryItemV2Request,
     ): ResponseEntity<Any> {
-        val userId = requireUserId(authentication) ?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build()
+        val userId = authPrincipalResolver.userId(authentication)
+            ?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build()
         val id = TripId.fromString(tripId) ?: return ResponseEntity.status(HttpStatus.BAD_REQUEST).build()
         val trip = itineraryV2Service.moveItem(
             tripId = id,
@@ -101,7 +102,7 @@ class ItineraryController(
             beforeItemId = request.beforeItemId,
             afterItemId = request.afterItemId,
         ) ?: return ResponseEntity.status(HttpStatus.NOT_FOUND).build()
-        return ResponseEntity.ok(toItineraryV2Response(trip))
+        return ResponseEntity.ok(ItineraryResponseMapper.toV2Response(trip))
     }
 
     @DeleteMapping("/v2/items/{itemId}")
@@ -110,56 +111,14 @@ class ItineraryController(
         @PathVariable tripId: String,
         @PathVariable itemId: String,
     ): ResponseEntity<Any> {
-        val userId = requireUserId(authentication) ?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build()
+        val userId = authPrincipalResolver.userId(authentication)
+            ?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build()
         val id = TripId.fromString(tripId) ?: return ResponseEntity.status(HttpStatus.BAD_REQUEST).build()
         val trip = itineraryV2Service.removeItem(
             tripId = id,
             userId = userId,
             itemId = itemId,
         ) ?: return ResponseEntity.status(HttpStatus.NOT_FOUND).build()
-        return ResponseEntity.ok(toItineraryV2Response(trip))
-    }
-
-    private fun requireUserId(authentication: Authentication): UserId? {
-        val principal = authentication.principal as? String ?: return null
-        return UserId.fromString(principal)
-    }
-
-    private fun toItineraryV2Response(trip: Trip): ItineraryV2Response {
-        val days = trip.generatedDays().map { day ->
-            DayContainerResponse(
-                dayNumber = day.dayNumber,
-                date = day.date.toString(),
-                items = day.items.map { item ->
-                    ItineraryItemV2Response(
-                        id = item.id.toString(),
-                        placeName = item.placeName,
-                        notes = item.notes,
-                        latitude = item.latitude,
-                        longitude = item.longitude,
-                        dayNumber = day.dayNumber,
-                    )
-                },
-            )
-        }
-
-        val places = ItemContainerResponse(
-            label = "Places To Visit",
-            items = trip.placesToVisitItems().map { item ->
-                ItineraryItemV2Response(
-                    id = item.id.toString(),
-                    placeName = item.placeName,
-                    notes = item.notes,
-                    latitude = item.latitude,
-                    longitude = item.longitude,
-                    dayNumber = null,
-                )
-            },
-        )
-
-        return ItineraryV2Response(
-            days = days,
-            placesToVisit = places,
-        )
+        return ResponseEntity.ok(ItineraryResponseMapper.toV2Response(trip))
     }
 }
